requirements.txt
absl-py==1.0.0
aiofiles==0.8.0
amqp==5.1.0
anyio==3.4.0
asgiref==3.4.1
astunparse==1.6.3
attrs==21.2.0
Automat==20.2.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
billiard==3.6.4.0
Brotli==1.0.9
bs4==0.0.1
cachetools==4.2.4
celery==5.2.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==36.0.0
cycler==0.11.0
dash==2.0.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.5.1
decorator==5.1.0
Deprecated==1.2.13
dicttoxml==1.7.4
ecdsa==0.17.0
entrypoints==0.3
fake-useragent==0.1.11
fastapi==0.70.1
Flask==2.0.2
Flask-Compress==1.10.1
flatbuffers==2.0
fonttools==4.28.3
gast==0.4.0
git-filter-repo==2.34.0
google-api-core==2.3.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-cloud-core==2.2.1
google-cloud-storage==1.43.0
google-crc32c==1.3.0
google-pasta==0.2.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
grpcio==1.42.0
h11==0.12.0
h5py==3.6.0
httptools==0.3.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.8.2
incremental==21.3.0
ipykernel==6.6.0
ipython==7.30.1
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
jupyter-client==7.1.0
jupyter-core==4.9.1
jwcrypto==1.0
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
kombu==5.2.4
libclang==12.0.0
lxml==4.7.1
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.4
numpy==1.21.4
oauthlib==3.1.1
opencv-python==4.5.4.60
opt-einsum==3.3.0
orjson==3.6.5
packaging==21.3
pandas==1.3.5
parso==0.8.3
passlib==1.7.4
pdf-tools==0.2.0
pdfkit==1.0.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
plotly==5.4.0
prompt-toolkit==3.0.24
protobuf==3.19.1
psycopg2-binary==2.9.2
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.8.2
Pygments==2.10.0
pymongo==4.0.1
PyMuPDF==1.19.4
pyparsing==3.0.6
python-dateutil==2.8.2
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
pytz==2021.3
PyYAML==6.0
pyzmq==22.3.0
qrcode==7.3.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
starlette==0.16.0
tenacity==8.0.1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.23.0
termcolor==1.1.0
tika==1.24
tornado==6.1
traitlets==5.1.1
Twisted==21.7.0
typing-extensions==4.0.1
urllib3==1.26.7
uvicorn==0.16.0
uvloop==0.16.0
vine==5.0.0
watchgod==0.7
wcwidth==0.2.5
websockets==10.1
Werkzeug==2.0.2
wkhtmltopdf==0.2
wrapt==1.13.3
xmltodict==0.12.0
zipp==3.6.0
zope.interface==5.4.0
billiard==3.6.4.0
